Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main() {
-    int radius,area;
+    float radius,area;
     scanf("%f",&radius);
     area=3.14*radius*radius;
     printf("Area: %.2f",area);
